fix: example components removed, file and component snames changed
